+++ lib/cache.dart
@@ -17,5 +17,7 @@ library quiver.cache;
 import 'dart:async';
 import 'dart:collection';
 
+import 'package:quiver/collection.dart' show LruMap;
+
 part 'src/cache/cache.dart';
 part 'src/cache/map_cache.dart';