+++ lib/collection.dart
@@ -20,7 +20,11 @@ library quiver.collection;
 import 'dart:collection';
 import 'dart:math';
 
+import 'package:quiver/core.dart';
+import 'package:quiver/iterables.dart';
+
 part 'src/collection/bimap.dart';
+part 'src/collection/lru_map.dart';
 part 'src/collection/multimap.dart';
 part 'src/collection/treeset.dart';
 part 'src/collection/delegates/iterable.dart';